+++ macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/opt/homebrew/Caskroom/flutter/3.22.0/flutter"
+export "FLUTTER_ROOT=/opt/homebrew/Caskroom/flutter/3.24.0/flutter"
 export "FLUTTER_APPLICATION_PATH=/Volumes/Data/Private/04_Librarys/simple_widget_snapshot"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=1.0.3"
-export "FLUTTER_BUILD_NUMBER=1.0.3"
+export "FLUTTER_BUILD_NAME=2.0.0"
+export "FLUTTER_BUILD_NUMBER=2.0.0"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"